Non-modal detail screen (#4)
Added a "detailAsModalShouldStayInPlace" option, which makes detail screen always open in-place instead of a separate modal screen, while hiding the tab bar and the master screens, if there's no horizontal space. 

The reason for introducing this feature is that, while you can open a 2nd level modal screen from a 1st level modal screen on iPad in 1/3 split screen, the app crashes if you try to change the split screen mode to a one, where the detail screen is not modal anymore.